+++ python/kpu/__init__.py
@@ -77,7 +77,7 @@ def _install_cpp_warning_filter():
 
 _flush_stderr = _install_cpp_warning_filter()
 
-__version__ = "0.6.3"
+__version__ = "0.6.4"
 __author__ = "Stillwater Supercomputing, Inc."
 
 # Fidelity levels